Replace misguided use of generated functions with regular functions
It was nice to be able to directly interpolate in many constants which
only depend on the Nside, but it was really hard to produce performant
user-facing code. Type instabilities occurred at any boundary between a
regular nside variable to Val{nside} type param, and these were
disastrous to loop performance.