Make use of fast sparse outer products from Julia (closes #4)
The feature was added to Julia in time for v1.2 in JuliaLang/julia#24980,
so get rid of the custom `outer()` method here and rewrite `quadprod()`
in terms of just standard matrix methods. Julia v1.2 is the
minimum-supported version at this point, so no need to worry about
backporting the functionality.

In the future, this function may yet still go away since the
implementation is nearly trivial at this point, but that can be a
follow-up PR.